+++ examples/pptxgenjs-demo.js
@@ -4,7 +4,7 @@
 * DESC: Common test/demo slides for all library features
 * DEPS: Loaded by `pptxgenjs-demo.js` and `nodejs-demo.js`
 * VER.: 2.5.0
-* BLD.: 20181202
+* BLD.: 20181230
 */
 
 // Detect Node.js (NODEJS is ultimately used to determine how to save: either `fs` or web-based, so using fs-detection is perfect)
@@ -1517,6 +1517,10 @@ function genSlides_Chart(pptx) {
 			valAxisOrientation: 'maxMin',
 			showCatAxisTitle: false,
 			showValAxisTitle: false,
+			catAxisMinVal: 0,
+			catAxisMaxVal: 4,
+			catAxisTitle: "Using catAxisMinVal/catAxisMaxVal",
+			showCatAxisTitle: true,
 			lineSize: 0
 		};
 		slide.addChart( pptx.charts.SCATTER, arrDataScatter1, optsChartScat3 );